Resumable image downloads

+++ app/src/main/java/eu/kanade/tachiyomi/data/download/Downloader.kt
@@ -40,6 +40,7 @@ import kotlinx.coroutines.supervisorScope
 import logcat.LogPriority
 import mihon.core.archive.ZipWriter
 import nl.adaptivity.xmlutil.serialization.XML
+import okhttp3.HttpUrl.Companion.toHttpUrl
 import okhttp3.Response
 import tachiyomi.core.common.i18n.stringResource
 import tachiyomi.core.common.storage.extension
@@ -361,11 +362,6 @@ class Downloader(
                 reIndexedPages
             }
 
-            // Delete all temporary (unfinished) files
-            tmpDir.listFiles()
-                ?.filter { it.extension == "tmp" }
-                ?.forEach { it.delete() }
-
             download.status = Download.State.DOWNLOADING
 
             // Start downloading images, consider we can have downloaded images already
@@ -440,14 +436,11 @@ class Downloader(
 
         val digitCount = (download.pages?.size ?: 0).toString().length.coerceAtLeast(3)
         val filename = "%0${digitCount}d".format(Locale.ENGLISH, page.number)
-        val tmpFile = tmpDir.findFile("$filename.tmp")
-
-        // Delete temp file if it exists
-        tmpFile?.delete()
 
         // Try to find the image file
         val imageFile = tmpDir.listFiles()?.firstOrNull {
-            it.name!!.startsWith("$filename.") || it.name!!.startsWith("${filename}__001")
+            (it.name!!.startsWith("$filename.") || it.name!!.startsWith("${filename}__001")) &&
+            !it.name!!.endsWith(".tmp")
         }
 
         try {
@@ -487,15 +480,25 @@ class Downloader(
         page.status = Page.State.DownloadImage
         page.progress = 0
         return flow {
+            val file = tmpDir.findFile("$filename.tmp")
+                ?: tmpDir.createFile("$filename.tmp")!!
+
+            page.imageUrl = page.imageUrl!!.toHttpUrl().newBuilder()
+                .addQueryParameter("existingSize", file.length().toString())
+                .build().toString()
+
             val response = source.getImage(page)
-            val file = tmpDir.createFile("$filename.tmp")!!
+
             try {
-                response.body.source().saveTo(file.openOutputStream())
+                // append if partial (HTTP 206), else overwrite
+                response.body.source().saveTo(file.openOutputStream(response.code == 206))
                 val extension = getImageExtension(response, file)
                 file.renameTo("$filename.$extension")
             } catch (e: Exception) {
                 response.close()
-                file.delete()
+                if (response.code == 416) {
+                    file.delete()
+                }
                 throw e
             }
             emit(file)
@@ -520,6 +523,8 @@ class Downloader(
      * @param filename the filename of the image.
      */
     private fun copyImageFromCache(cacheFile: File, tmpDir: UniFile, filename: String): UniFile {
+        // Delete temp file if it exists
+        tmpDir.findFile("$filename.tmp")?.delete()
         val tmpFile = tmpDir.createFile("$filename.tmp")!!
         cacheFile.inputStream().use { input ->
             tmpFile.openOutputStream().use { output ->
